+++ asyncmy/connection.pyx
@@ -11,7 +11,6 @@ import warnings
 from asyncio import StreamReader, StreamWriter
 from typing import Optional, Type
 
-import xstruct as struct
 from asyncmy import auth, converters, errors
 from asyncmy.charset import charset_by_id, charset_by_name
 from asyncmy.cursors import Cursor
@@ -36,6 +35,7 @@ from .constants.SERVER_STATUS import (SERVER_STATUS_AUTOCOMMIT,
                                       SERVER_STATUS_IN_TRANS,
                                       SERVER_STATUS_NO_BACKSLASH_ESCAPES)
 from .contexts import _ConnectionContextManager
+from .structs import B_, BHHB, HBB, IIB, B, H, I, Q, i, iB, iIB23s
 from .version import __VERSION__
 
 try:
@@ -72,7 +72,7 @@ cdef str DEFAULT_CHARSET = "utf8mb4"
 cdef int MAX_PACKET_LEN = 2 ** 24 - 1
 
 cdef _pack_int24(int n):
-    return struct.pack("<I", n)[:3]
+    return I.pack(n)[:3]
 
 # https://dev.mysql.com/doc/internals/en/integer.html#packet-Protocol::LengthEncodedInteger
 cdef _lenenc_int(int i):
@@ -83,11 +83,11 @@ cdef _lenenc_int(int i):
     elif i < 0xFB:
         return bytes([i])
     elif i < (1 << 16):
-        return b"\xfc" + struct.pack("<H", i)
+        return b"\xfc" + H.pack(i)
     elif i < (1 << 24):
-        return b"\xfd" + struct.pack("<I", i)[:3]
+        return b"\xfd" + I.pack(i)[:3]
     elif i < (1 << 64):
-        return b"\xfe" + struct.pack("<Q", i)
+        return b"\xfe" + Q.pack(i)
     else:
         raise ValueError(
             "Encoding %x is larger than %x - no representation in LengthEncodedInteger"
@@ -177,11 +177,10 @@ class Connection:
             program_name=None,
             server_public_key=None,
             echo=False,
-            loop=None,
             ssl=None,
             db=None,  # deprecated
     ):
-        self._loop = loop or asyncio.get_event_loop()
+        self._loop = asyncio.get_event_loop()
         self._last_usage = self._loop.time()
         if db is not None and database is None:
             # We will raise warining in 2022 or later.
@@ -316,7 +315,7 @@ class Connection:
     async def ensure_closed(self):
         """Close connection without QUIT message."""
         if self._connected:
-            send_data = struct.pack('<i', 1) + struct.pack("!B", 0x01)  # for cython build
+            send_data = i.pack(1) + B.pack(COM_QUIT)
             self._write_bytes(send_data)
             await self._writer.drain()
             self._writer.close()
@@ -463,8 +462,8 @@ class Connection:
     def affected_rows(self):
         return self._affected_rows
 
-    async def kill(self, int thread_id):
-        arg = struct.pack("<I", thread_id)
+    async def kill(self, thread_id):
+        arg = I.pack(thread_id)
         await self._execute_command(COM_PROCESS_KILL, arg)
         return await self._read_ok_packet()
 
@@ -509,8 +508,7 @@ class Connection:
 
             if self._unix_socket:
                 self._reader, self._writer = await asyncio.wait_for(asyncio.open_unix_connection(self._unix_socket),
-                                                                    timeout=self._connect_timeout,
-                                                                    loop=self._loop)
+                                                                    timeout=self._connect_timeout,)
                 self.host_info = "Localhost via UNIX socket"
                 self._secure = True
             else:
@@ -519,7 +517,6 @@ class Connection:
                         self._reader, self._writer = await asyncio.wait_for(asyncio.open_connection(
                             self._host,
                             self._port,
-                            loop=self._loop,
                         ), timeout=self._connect_timeout)
                         self._set_keep_alive()
                         break
@@ -564,7 +561,7 @@ class Connection:
         """
         # Internal note: when you build packet manually and calls _write_bytes()
         # directly, you should set self._next_seq_id properly.
-        data = _pack_int24(len(payload)) + struct.pack("!B", self._next_seq_id) + payload
+        data = _pack_int24(len(payload)) + B.pack(self._next_seq_id) + payload
         self._write_bytes(data)
         self._next_seq_id = (self._next_seq_id + 1) % 256
 
@@ -578,8 +575,8 @@ class Connection:
         """
         buff = bytearray()
         while True:
-            packet_header: bytes = await self._read_bytes(4)
-            btrl, btrh, packet_number = struct.unpack("<HBB", packet_header)
+            packet_header = await self._read_bytes(4)
+            btrl, btrh, packet_number = HBB.unpack(packet_header)
             bytes_to_read = btrl + (btrh << 16)
             if packet_number != self._next_seq_id:
                 if packet_number == 0:
@@ -657,7 +654,7 @@ class Connection:
         else:
             await self.ensure_closed()
 
-    async def _execute_command(self, bytes command, sql):
+    async def _execute_command(self, command, sql):
         """
         :raise InterfaceError: If the connection is closed.
         :raise ValueError: If no username was specified.
@@ -678,11 +675,11 @@ class Connection:
         if isinstance(sql, str):
             sql = sql.encode(self._encoding)
 
-        cdef int packet_size = min(MAX_PACKET_LEN, len(sql) + 1)  # +1 is for command
+        packet_size = min(MAX_PACKET_LEN, len(sql) + 1)  # +1 is for command
 
         # tiny optimization: build first packet manually instead of
         # calling self..write_packet()
-        prelude = struct.pack("<iB", packet_size, command)
+        prelude = iB.pack(packet_size, command)
         self._write_bytes(prelude + sql[: packet_size - 1])
         self._next_seq_id = 1
 
@@ -707,7 +704,7 @@ class Connection:
 
         if self._ssl_context:
             # capablities, max packet, charset
-            data = struct.pack('<IIB', self._client_flag, 16777216, 33)
+            data = IIB.pack('<IIB', self._client_flag, 16777216, 33)
             data += b'\x00' * (32 - len(data))
 
             self.write_packet(data)
@@ -729,13 +726,13 @@ class Connection:
             # connection not initiate a new one.
             self._reader, self._writer = await asyncio.open_connection(
                 sock=raw_sock, ssl=self._ssl_context,
-                server_hostname=self._host, loop=self._loop,
+                server_hostname=self._host,
             )
-        cdef int charset_id = charset_by_name(self._charset).id
+        charset_id = charset_by_name(self._charset).id
         if isinstance(self._user, str):
             self._user = self._user.encode(self._encoding)
 
-        data_init = struct.pack("<iIB23s", self._client_flag, MAX_PACKET_LEN, charset_id, b"")
+        data_init = iIB23s.pack(self._client_flag, MAX_PACKET_LEN, charset_id, b"")
         data = data_init + self._user + b"\0"
 
         authresp = b""
@@ -763,7 +760,7 @@ class Connection:
         if self.server_capabilities & PLUGIN_AUTH_LENENC_CLIENT_DATA:
             data += _lenenc_int(len(authresp)) + authresp
         elif self.server_capabilities & SECURE_CONNECTION:
-            data += struct.pack("B", len(authresp)) + authresp
+            data += B_.pack(len(authresp)) + authresp
         else:  # pragma: no cover - not testing against servers without secure auth (>=5.0)
             data += authresp + b"\0"
 
@@ -779,10 +776,10 @@ class Connection:
             connect_attrs = b""
             for k, v in self._connect_attrs.items():
                 k = k.encode("utf-8")
-                connect_attrs += struct.pack("B", len(k)) + k
+                connect_attrs += B_.pack(len(k)) + k
                 v = v.encode("utf-8")
-                connect_attrs += struct.pack("B", len(v)) + v
-            data += struct.pack("B", len(connect_attrs)) + connect_attrs
+                connect_attrs += B_.pack(len(v)) + v
+            data += B_.pack(len(connect_attrs)) + connect_attrs
 
         self.write_packet(data)
         auth_packet = await self.read_packet()
@@ -923,7 +920,7 @@ class Connection:
     async def _get_server_information(self):
         i = 0
         packet = await self.read_packet()
-        data: bytes = packet.get_all_data()
+        data = packet.get_all_data()
 
         self.protocol_version = data[i]
         i += 1
@@ -932,17 +929,17 @@ class Connection:
         self.server_version = data[i:server_end].decode("latin1")
         i = server_end + 1
 
-        self.server_thread_id = struct.unpack("<I", data[i: i + 4])
+        self.server_thread_id = I.unpack(data[i: i + 4])
         i += 4
 
         self.salt = data[i: i + 8]
         i += 9  # 8 + 1(filler)
 
-        self.server_capabilities = struct.unpack("<H", data[i: i + 2])[0]
+        self.server_capabilities = H.unpack(data[i: i + 2])[0]
         i += 2
 
         if len(data) >= i + 6:
-            lang, stat, cap_h, salt_len = struct.unpack("<BHHB", data[i: i + 6])
+            lang, stat, cap_h, salt_len = BHHB.unpack(data[i: i + 6])
             i += 6
             # TODO: deprecate server_language and server_charset.
             # mysqlclient-python doesn't provide it.
@@ -1153,7 +1150,7 @@ cdef class MySQLResult:
                 # No more columns in this row
                 # See https://github.com/PyMySQL/PyMySQL/pull/434
                 break
-            if data != b"":
+            if data is not None:
                 if encoding is not None:
                     data = data.decode(encoding)
                 if converter is not None:
@@ -1256,7 +1253,6 @@ def connect(user=None,
             write_timeout=None,
             binary_prefix=False,
             program_name=None,
-            loop=None,
             echo=False,
             server_public_key=None,
             ssl=None,
@@ -1287,7 +1283,6 @@ def connect(user=None,
         write_timeout=write_timeout,
         binary_prefix=binary_prefix,
         program_name=program_name,
-        loop=loop,
         server_public_key=server_public_key,
         echo=echo,
         ssl=ssl,